reorganizing the repo
move the go files to pkg folder
