+++ Dockerfile
@@ -15,8 +15,7 @@ RUN ls
 
 # Copy the go source
 COPY main.go main.go
-COPY api/ api/
-COPY controllers/ controllers/
+COPY pkg/ pkg/
 COPY charts/ charts/
 
 # Build
