+++ pkg/azure/utils.go
