+++ pkg/controllers/egressrules_controller.go
@@ -19,8 +19,8 @@ package controllers
 import (
 	"context"
 
-	egressv1 "azure-firewall-egress-controller.io/aks-egress/api/v1"
-	a "azure-firewall-egress-controller.io/aks-egress/azure"
+	egressv1 "azure-firewall-egress-controller.io/aks-egress/pkg/api/v1"
+	a "azure-firewall-egress-controller.io/aks-egress/pkg/azure"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
